refactor: split monolithic server.js into server/ modules
