+++ Dockerfile
@@ -41,6 +41,7 @@ COPY --from=build /app/src/active-branch.json ./active-branch.json
 COPY package*.json ./
 RUN npm install --production
 COPY server.js .
+COPY server/ ./server/
 COPY database.json .
 COPY migrations/ ./migrations/
