+++ package.json
@@ -57,7 +57,7 @@
     "serve": "node server.js",
     "migrate": "db-migrate up",
     "watch:build": "nodemon --watch src --ext js,jsx,css --exec \"npm run build\" --ignore build",
-    "watch:server": "nodemon --watch server.js --exec \"node server.js\" --ignore build --ignore src",
+    "watch:server": "nodemon --watch server.js --watch server --exec \"node server.js\" --ignore build --ignore src",
     "dev": "concurrently \"npm run watch:build\" \"npm run watch:server\" --names \"BUILD,SERVER\" --prefix-colors \"cyan,yellow\""
   },
   "browserslist": {
